Move all `theme` function backwards compatibility concerns to compat layer
