+++ packages/tailwindcss/src/theme.ts
@@ -54,7 +54,7 @@ export class Theme {
     return keys
   }
 
-  get(themeKeys: (ThemeKey | `${ThemeKey}-${string}`)[]): string | null {
+  get(themeKeys: ThemeKey[]): string | null {
     for (let key of themeKeys) {
       let value = this.values.get(key)
       if (value) {
@@ -179,130 +179,4 @@ export class Theme {
   }
 }
 
-export type ThemeKey =
-  | '--accent-color'
-  | '--animate'
-  | '--aspect-ratio'
-  | '--backdrop-blur'
-  | '--backdrop-brightness'
-  | '--backdrop-contrast'
-  | '--backdrop-grayscale'
-  | '--backdrop-hue-rotate'
-  | '--backdrop-invert'
-  | '--backdrop-opacity'
-  | '--backdrop-saturate'
-  | '--backdrop-sepia'
-  | '--background-color'
-  | '--background-image'
-  | '--blur'
-  | '--border-color'
-  | '--border-spacing'
-  | '--border-width'
-  | '--box-shadow-color'
-  | '--breakpoint'
-  | '--brightness'
-  | '--caret-color'
-  | '--color'
-  | '--columns'
-  | '--contrast'
-  | '--cursor'
-  | '--default-border-width'
-  | '--default-ring-color'
-  | '--default-ring-width'
-  | '--default-transition-timing-function'
-  | '--default-transition-duration'
-  | '--divide-width'
-  | '--divide-color'
-  | '--drop-shadow'
-  | '--fill'
-  | '--flex-basis'
-  | '--font-family'
-  | '--font-size'
-  | '--font-weight'
-  | '--gap'
-  | '--gradient-color-stop-positions'
-  | '--grayscale'
-  | '--grid-auto-columns'
-  | '--grid-auto-rows'
-  | '--grid-column'
-  | '--grid-column-end'
-  | '--grid-column-start'
-  | '--grid-row'
-  | '--grid-row-end'
-  | '--grid-row-start'
-  | '--grid-template-columns'
-  | '--grid-template-rows'
-  | '--height'
-  | '--hue-rotate'
-  | '--inset'
-  | '--inset-shadow'
-  | '--invert'
-  | '--letter-spacing'
-  | '--line-height'
-  | '--line-clamp'
-  | '--list-style-image'
-  | '--list-style-type'
-  | '--margin'
-  | '--max-height'
-  | '--max-width'
-  | '--min-height'
-  | '--min-width'
-  | '--object-position'
-  | '--opacity'
-  | '--order'
-  | '--outline-color'
-  | '--outline-width'
-  | '--outline-offset'
-  | '--padding'
-  | '--placeholder-color'
-  | '--perspective'
-  | '--perspective-origin'
-  | '--radius'
-  | '--ring-color'
-  | '--ring-offset-color'
-  | '--ring-offset-width'
-  | '--ring-width'
-  | '--rotate'
-  | '--saturate'
-  | '--scale'
-  | '--scroll-margin'
-  | '--scroll-padding'
-  | '--sepia'
-  | '--shadow'
-  | '--size'
-  | '--skew'
-  | '--space'
-  | '--spacing'
-  | '--stroke'
-  | '--stroke-width'
-  | '--text-color'
-  | '--text-decoration-color'
-  | '--text-decoration-thickness'
-  | '--text-indent'
-  | '--text-underline-offset'
-  | '--transform-origin'
-  | '--transition-delay'
-  | '--transition-duration'
-  | '--transition-property'
-  | '--transition-timing-function'
-  | '--translate'
-  | '--width'
-  | '--z-index'
-  | `--default-${string}`
-
-export type ColorThemeKey =
-  | '--color'
-  | '--accent-color'
-  | '--background-color'
-  | '--border-color'
-  | '--box-shadow-color'
-  | '--caret-color'
-  | '--divide-color'
-  | '--fill'
-  | '--outline-color'
-  | '--placeholder-color'
-  | '--ring-color'
-  | '--ring-offset-color'
-  | '--stroke'
-  | '--text-color'
-  | '--text-decoration-color'
+export type ThemeKey = `--${string}`
